chore: relocate unshared tx warning styles
